# llm-gateway-contract/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: LLM Gateway Public Client API
  version: 0.1.0
  description: >-
    Provider-neutral chat completion contract for BFF and domain backend clients.
    The internal routing control plane is deliberately excluded from this public contract.
servers:
  - url: http://localhost:8080
paths:
  /v1/chat/completions:
    post:
      operationId: createChatCompletion
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Completed JSON response or Server-Sent Events when stream is true. After an SSE response starts, errors are delivered as an event named error.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
            text/event-stream:
              schema:
                type: string
                description: >-
                  Normal events contain a JSON ChatCompletionChunk. Error events use event: error and contain ErrorResponse.
                  The success terminal and data: [DONE] follow successful accounting writes and stream cleanup.
                  A storage failure ends with OUTCOME_UNKNOWN, retryable=false, without a success terminal or DONE.
                  Earlier content cannot be retracted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/GatewayError'
        '503':
          description: Gateway unavailable, or OUTCOME_UNKNOWN if durable completion could not be confirmed. OUTCOME_UNKNOWN is not retryable and has no Retry-After header; disable SDK automatic retries for generation requests.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  type: outcome_unknown
                  code: OUTCOME_UNKNOWN
                  message: The gateway could not confirm durable completion; do not repeat this generation request
                  retryable: false
                  request_id: req-example
        '504':
          $ref: '#/components/responses/GatewayTimeout'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Optional caller-supplied correlation id. The gateway always returns the effective id.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
    Traceparent:
      name: traceparent
      in: header
      required: false
      description: Optional W3C trace context used to correlate gateway observations.
      schema:
        type: string
        pattern: '^[0-9A-Fa-f]{2}-[0-9A-Fa-f]{32}-[0-9A-Fa-f]{16}-[0-9A-Fa-f]{2}$'
  headers:
    RequestId:
      description: Effective request correlation id.
      schema:
        type: string
    RetryAfter:
      description: Number of seconds before retrying the request.
      schema:
        type: integer
        format: int64
        minimum: 0
  responses:
    BadRequest:
      description: Caller-fixable request error.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid gateway credential.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated client is not allowed to use the requested gateway capability.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: The request or model response was rejected by a gateway policy.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: >-
        Gateway/provider rate limit or BUDGET_EXCEEDED. A budget that requires
        administrative funding returns retryable=false without Retry-After;
        clients must not infer an automatic reset or retry from HTTP 429 alone.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    GatewayError:
      description: Gateway or provider failure.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    GatewayTimeout:
      description: The gateway deadline expired before a complete response was produced.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ChatCompletionRequest:
      type: object
      additionalProperties: false
      required: [model, messages]
      properties:
        model:
          type: string
          minLength: 1
          description: Logical model group, not a provider-specific model id.
        messages:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ChatMessage'
        temperature:
          type: number
          minimum: 0
          maximum: 2
        max_tokens:
          type: integer
          minimum: 1
          description: Backward-compatible alias for max_completion_tokens. Do not send both.
        max_completion_tokens:
          type: integer
          minimum: 1
          description: Maximum number of tokens to generate. Do not send with max_tokens.
        top_p:
          type: number
          minimum: 0
          maximum: 1
        stop:
          type: array
          items:
            type: string
        stream:
          type: boolean
          default: false
          description: When true, the response is text/event-stream. This flag controls streaming.
    ChatMessage:
      type: object
      additionalProperties: false
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string
    ChatCompletionResponse:
      type: object
      additionalProperties: false
      required: [id, object, created, model, choices]
      properties:
        id:
          type: string
        object:
          type: string
          enum: [chat.completion]
        created:
          type: integer
          format: int64
        model:
          type: string
          description: Logical model group requested by the client.
        choices:
          minItems: 1
          maxItems: 1
          type: array
          items:
            $ref: '#/components/schemas/ChatChoice'
        usage:
          description: Provider usage metadata; omitted when the provider does not return usage data.
          allOf:
            - $ref: '#/components/schemas/Usage'
    ChatCompletionChunk:
      type: object
      additionalProperties: false
      required: [id, object, created, model, choices]
      properties:
        id: { type: string }
        object: { type: string, enum: [chat.completion.chunk] }
        created: { type: integer, format: int64 }
        model:
          type: string
          description: Logical model group requested by the client.
        choices:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/ChatChunkChoice'
        usage:
          description: Provider usage metadata when available; omitted otherwise.
          allOf:
            - $ref: '#/components/schemas/Usage'
    ChatChoice:
      type: object
      additionalProperties: false
      required: [index, message]
      properties:
        index: { type: integer }
        message: { $ref: '#/components/schemas/ChatMessage' }
        finish_reason: { type: string, nullable: true }
    ChatChunkChoice:
      type: object
      additionalProperties: false
      required: [index, delta]
      properties:
        index: { type: integer }
        delta: { $ref: '#/components/schemas/ChatDelta' }
        finish_reason: { type: string, nullable: true }
    Usage:
      type: object
      additionalProperties: false
      required: [prompt_tokens, completion_tokens, total_tokens]
      properties:
        prompt_tokens: { type: integer, format: int64, minimum: 0 }
        completion_tokens: { type: integer, format: int64, minimum: 0 }
        total_tokens: { type: integer, format: int64, minimum: 0 }
    ChatDelta:
      type: object
      additionalProperties: false
      properties:
        role:
          type: string
          enum: [assistant]
        content:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/GatewayError' }
    GatewayError:
      type: object
      additionalProperties: false
      required: [type, code, message, retryable, request_id]
      properties:
        type: { type: string }
        code: { type: string }
        message: { type: string }
        retryable: { type: boolean }
        request_id: { type: string }
        retry_after_seconds: { type: integer, format: int64, minimum: 0 }